[multiplex] OAK-3403 - Multiplexing store support in Property indexes
Partial support for unique indexes. This would fail in cluster setup is two unique paths with same key are inserted  at same time in different stores

With this both count and query methods are not multiplexed

exists call
------------

The exists call is only supported by Unique strategy. To get it working properly have to break the abstraction a bit as PropertyIndexEditor would now need to call a method specific to MultiplexingStrategy